% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_methods.R
\docType{methods}
\name{show,powerLongSurv-method}
\alias{show,powerLongSurv-method}
\title{Methods for Function \code{show}}
\usage{
\S4method{show}{powerLongSurv}(object)
}
\arguments{
\item{object}{An object of class \code{powerLongSurv}}
}
\description{
Methods for function \code{show}
}
\keyword{methods}
